text uses a new referencing convention
The text shows a very easy way for authors to reference to terms. For example, suppose we operate under the scope `foo`, and we want to make some text `bars` refer (hyperlink) to the term:

- `bar` as defined in scope `foo`. The syntax is `[bars](bar)`.
- `bar` as defined in scope `footon`. The syntax is then `[bars](bar@footon)`.

Appropriate postprocessing needs to ensure that the links `bar` and `bar@footon` are replaced by the appropriate hyperlinks.